requirements.txt
asn1crypto==0.24.0
blinker==1.4
chardet==3.0.4
cliapp==1.20170827
Click==7.0
cmdtest==0.32
cryptography==2.1.4
dominate==2.3.5
enum34==1.1.6
Flask==1.0.2
Flask-Bootstrap4==4.0.2
Flask-Mail==0.9.1
Flask-Uploads==0.2.1
idna==2.6
ipaddress==1.0.17
itsdangerous==1.1.0
Jinja2==2.10
keyring==10.6.0
keyrings.alt==3.0
Markdown==2.6.9
MarkupSafe==1.1.0
pycrypto==2.6.1
Pygments==2.2.0
pygobject==3.26.1
pyxdg==0.25
PyYAML==3.12
SecretStorage==2.3.1
six==1.11.0
ttystatus==0.38
visitor==0.1.3
Werkzeug==0.14.1
